# Avoid a copy when optimizing part size for checksums