## Description of change

The `chunks` iterator returns slices, so creating the `Bytes` for each
chunk needs to do a copy. We can instead just do the `Bytes` conversion
upfront from the `Box<[u8]>`, and then do O(1) splits of that `Bytes`
object for each chunk.

Before:

```
$ time dd if=$HOME/mnt/test100GiB-STD of=/dev/null bs=1M iflag=direct status=progress
107374182400 bytes (107 GB, 100 GiB) copied, 38.7547 s, 2.8 GB/s
```

After:
```
$ time dd if=$HOME/mnt/test100GiB-STD of=/dev/null bs=1M iflag=direct status=progress
107374182400 bytes (107 GB, 100 GiB) copied, 33.9263 s, 3.2 GB/s
```

## Does this change impact existing behavior?

No, this is an internal performance improvement.

---

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license and I agree to the terms of the [Developer Certificate of Origin (DCO)](https://developercertificate.org/).